Prevent specific while loop to be recognized as heredoc
This change should fix issues with while loops formatted the following way:
  while read -r thing; do
    echo "$thing done"
  done <<< $filename

beautysh.py was parsing that `<<< $filename` as a heredoc delimiter, enabling a heredoc paragraph that was never being closed.
And that was the cause for the error message of those indent/outdent mismatch: the indentation was the same as the one in `done <<< $filename` till the end of file.
Now, before enabling heredoc, we double check if it really isn't a while case.

Ref: https://github.com/bemeurer/beautysh/issues/23